requirements.txt
Django==5.0.4
black==24.4.0
isort==5.13.2
flake8==7.0.0
djangorestframework==3.15.1
django-configurations==2.5.1
djangorestframework-simplejwt==5.3.1
